+++ docs/CHANGELOG.md
@@ -1,3 +1,12 @@
+# Next
+
+* Optional preloading CPU PJRT plugin:
+  * `github.com/gomlx/gopjrt/pjrt/cpu/static` that statically links the PJRT CPU plugin: easy to deploy binary. 
+    It includes the corresponding C BUILD rule to build the static library (`libpjrt_c_api_cpu_static.a`)
+  * `github.com/gomlx/gopjrt/pjrt/cpu/dynamic` that dynamically links (and preloads) the PJRT CPU plugin.
+* `pjrt_c_api_cpu.so` now compiled directly from `gopjrt`, and doesn't require cloning `xla` separately. It will
+  be distributed in the same `tar.gz` file.
+
 # v0.4.8 - 2024-11-19
 
 * Replaced C++ `xla::StatusOr` by `absl::StatusOr` (the former was already an alias to the later).